BGDIINF_SB-1799: Update Makefile, Readme, move functions into s3 and dynamodb
