+++ .isort.cfg
@@ -2,5 +2,7 @@
 known_third_party=pytest
 known_defusedxml=defusedxml
 known_flask=flask
+known_boto3=boto3
+known_botocore=botocore
 force_single_line=True
-sections=FUTURE,STDLIB,THIRDPARTY,DEFUSEDXML,FLASK,FIRSTPARTY,LOCALFOLDER
+sections=FUTURE,STDLIB,THIRDPARTY,DEFUSEDXML,FLASK,BOTO3,BOTOCORE,FIRSTPARTY,LOCALFOLDER
