+++ docker-compose.yml
@@ -1,4 +1,4 @@
-version: "2"
+version: "3.4"
 services:
   s3:
     image: minio/minio
@@ -38,7 +38,14 @@ services:
       - "8080:8000"
   init-dynamo:
     image: amazon/aws-cli
-    command: "dynamodb create-table --table-name test-db --attribute-definitions AttributeName=adminId,AttributeType=S --key-schema AttributeName=adminId,KeyType=HASH --provisioned-throughput ReadCapacityUnits=1,WriteCapacityUnits=1 --endpoint-url http://dynamodb-local:8000"
+    command: [
+      "dynamodb",
+      "create-table",
+      "--table-name test-db",
+      "--attribute-definitions AttributeName=file_id,AttributeType=S",
+      "--key-schema AttributeName=file_id,KeyType=HASH",
+      "--provisioned-throughput ReadCapacityUnits=1,WriteCapacityUnits=1",
+      "--endpoint-url http://dynamodb-local:8000"]
     links:
       - dynamodb-local
     environment:
